+++ packages/react-mui-audit-log-viewer/jest.config.js
@@ -1,0 +1,29 @@
+module.exports = {
+  collectCoverageFrom: [
+    "**/*.{js,jsx,ts,tsx}",
+    "!**/*.d.ts",
+    "!**/node_modules/**",
+    "!**/*.stories.tsx",
+  ],
+  moduleNameMapper: {
+    /* Handle CSS imports (with CSS modules)
+          https://jestjs.io/docs/webpack#mocking-css-modules */
+    "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy",
+
+    // Handle CSS imports (without CSS modules)
+    "^.+\\.(css|sass|scss)$": "<rootDir>/__mocks__/styleMock.js",
+
+    /* Handle image imports
+          https://jestjs.io/docs/webpack#handling-static-assets */
+    "^.+\\.(jpg|jpeg|png|gif|webp|avif|svg)$":
+      "<rootDir>/__mocks__/fileMock.js",
+  },
+  testPathIgnorePatterns: ["<rootDir>/node_modules/"],
+  transform: {
+    "^.+\\.(ts|tsx)$": ["ts-jest"],
+  },
+  transformIgnorePatterns: [
+    "/node_modules/",
+    "^.+\\.module\\.(css|sass|scss)$",
+  ],
+};