+++ packages/react-mui-audit-log-viewer/src/AuditLogViewer.tsx
@@ -20,6 +20,7 @@ export interface AuditLogViewerProps<Event = Audit.DefaultEvent> {
 
   onSearch: (body: Audit.SearchRequest) => Promise<Audit.SearchResponse>;
   searchOnChange?: boolean;
+  searchOnFilterChange?: boolean;
 
   onPageChange: (body: Audit.ResultRequest) => Promise<Audit.ResultResponse>;
   onDownload?: (
@@ -31,6 +32,11 @@ export interface AuditLogViewerProps<Event = Audit.DefaultEvent> {
     ModalChildComponent?: FC;
     onCopy?: (message: string, value: string) => void;
   };
+
+  fpeOptions?: {
+    highlightRedaction?: boolean;
+  };
+
   sx?: SxProps;
   pageSize?: number;
   dataGridProps?: Partial<DataGridProps>;
@@ -55,6 +61,7 @@ const AuditLogViewerWithProvider = <Event,>({
   schemaOptions,
   initialQuery,
   filters,
+  fpeOptions,
   ...props
 }: AuditLogViewerProps<Event>): JSX.Element => {
   const { schema } = useSchema(config, schemaProp, schemaOptions);
@@ -70,7 +77,12 @@ const AuditLogViewerWithProvider = <Event,>({
 
   const handleSearch = (body: Audit.SearchRequest): Promise<void> => {
     setLoading(true);
-    return onSearch(body)
+    return onSearch({
+      ...body,
+      ...(!!fpeOptions?.highlightRedaction && {
+        return_context: true,
+      }),
+    })
       .then((response) => {
         setLoading(false);
         if (!response || response?.events === undefined) {
@@ -99,7 +111,12 @@ const AuditLogViewerWithProvider = <Event,>({
 
   const handleResults = (body: Audit.ResultRequest): Promise<void> => {
     setLoading(true);
-    return onPageChange(body)
+    return onPageChange({
+      ...body,
+      ...(!!fpeOptions?.highlightRedaction && {
+        return_context: true,
+      }),
+    })
       .then((response) => {
         setLoading(false);
         if (!response) return;
@@ -120,7 +137,12 @@ const AuditLogViewerWithProvider = <Event,>({
     if (!onDownload) return;
 
     setLoading(true);
-    return onDownload(body)
+    return onDownload({
+      ...body,
+      ...(!!fpeOptions?.highlightRedaction && {
+        return_context: true,
+      }),
+    })
       .then((response) => {
         setLoading(false);
         if (response?.dest_url) {