+++ packages/react-mui-audit-log-viewer/src/components/AuditOldNewFields/index.tsx
@@ -1,4 +1,4 @@
-import { FC, useState } from "react";
+import { FC, useMemo, useState } from "react";
 import isEmpty from "lodash/isEmpty";
 
 import { Stack, Typography, Switch } from "@mui/material";
@@ -7,6 +7,12 @@ import { Audit } from "../../types";
 import { useDiffWords } from "../../hooks/diff";
 
 import StringJsonField from "../AuditStringJsonField";
+import {
+  FPEContext,
+  getFieldMatches,
+  injectFPEMatchesIntoChanges,
+  updateMatchesToHaveRelativeRanges,
+} from "../../hooks/fpe";
 
 interface Props {
   event: {
@@ -15,53 +21,71 @@ interface Props {
   };
   direction: "row" | "column";
   uniqueId: string;
+  context?: FPEContext;
 }
 
-const OldNewFields: FC<Props> = ({ event, direction, uniqueId }) => {
-  const [showDiff, setShowDiff] = useState(true);
+const OldNewFields: FC<Props> = ({ event, direction, uniqueId, context }) => {
   const changes = useDiffWords(event.old, event.new);
 
-  const shouldShowDiffs = showDiff && !isEmpty(changes);
+  const oldChanges = useMemo(() => {
+    const matches = getFieldMatches("old", context);
+    const changes_ = changes.filter((c) => !c.added);
+
+    try {
+      // Adjust field FPE matches to be relative to the entire strinified object
+      // to work with the used JSONViewer, which does not have support for highlighting
+      // values based on their JSON path
+      const adjustedMatches = updateMatchesToHaveRelativeRanges(
+        event.old ?? "",
+        matches
+      );
+      return injectFPEMatchesIntoChanges(adjustedMatches, changes_);
+    } catch {
+      // Unable to adjust field matches to be relative to complete string
+    }
+
+    return changes_;
+  }, []);
+
+  const newChanges = useMemo(() => {
+    const matches = getFieldMatches("new", context);
+    const changes_ = changes.filter((c) => !c.removed);
+
+    try {
+      // Adjust field FPE matches to be relative to the entire strinified object
+      // to work with the used JSONViewer, which does not have support for highlighting
+      // values based on their JSON path
+      const adjustedMatches = updateMatchesToHaveRelativeRanges(
+        event.new ?? "",
+        matches
+      );
+      return injectFPEMatchesIntoChanges(adjustedMatches, changes_);
+    } catch {
+      // Unable to adjust field matches to be relative to complete string
+    }
+
+    return changes_;
+  }, []);
 
   return (
     <Stack sx={{ position: "relative" }} spacing={-1}>
-      {!isEmpty(changes) && false && (
-        <Stack
-          direction="row"
-          spacing={1}
-          marginLeft="auto"
-          paddingRight={2}
-          sx={{
-            alignItems: "center",
-          }}
-        >
-          <Typography color="textSecondary" variant="body2">
-            {showDiff ? "Hide differences" : "Show differences"}
-          </Typography>
-          <Switch
-            checked={showDiff}
-            color="info"
-            onClick={() => setShowDiff(!showDiff)}
-          />
-        </Stack>
-      )}
       {!!event.old && (
         <StringJsonField
           title="Old Value"
           field="old"
           value={event.old}
-          shouldHighlight={(c) => !!c.removed}
-          changes={shouldShowDiffs ? changes.filter((c) => !c.added) : []}
+          shouldHighlight={(c) => !!c.removed || !!c.redacted}
+          changes={oldChanges}
           uniqueId={uniqueId}
         />
       )}
       {!!event.new && (
         <StringJsonField
           title="New Value"
-          field="old"
+          field="new"
           value={event.new}
-          shouldHighlight={(c) => !!c.added}
-          changes={shouldShowDiffs ? changes.filter((c) => !c.removed) : []}
+          shouldHighlight={(c) => !!c.added || !!c.redacted}
+          changes={newChanges}
           uniqueId={uniqueId}
         />
       )}